jets: move shared functions into jets::util(::test)
